requirements.txt
Flask
PyPDF2
langchain-community
langchain_ollama
langchain.text_splitter
langchain.chains
langchain.prompts